using get_bookings() instead of just bookings property
git-svn-id: https://plugins.svn.wordpress.org/events-manager/trunk@298896 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ admin/events.php
@@ -22,7 +22,7 @@ function dbem_events_subpanel() {
 	$event_table_name = $wpdb->prefix . EVENTS_TBNAME;	
 
 	// DELETE action
-	if ($action == 'deleteEvents') {
+	if ( $action == 'deleteEvents' && EM_Object::array_is_numeric($selectedEvents) ) {
 		EM_Events::delete( $selectedEvents );
 		dbem_events_table ( EM_Events::get( array('scope'=>$scope) ), "Future events" );
 	}
